Clean up before Snowflake otel e2e test.
